Fix calculation of maxRangeIntervalMs to avoid unnecessary casting

+++ kafka-connector-project/kafka-connector/src/main/java/com/lightstreamer/kafka/common/monitors/KafkaConnectorMonitor.java
@@ -200,7 +200,7 @@ public final class KafkaConnectorMonitor implements Monitor {
             if (interval == null || interval.isZero() || interval.isNegative()) {
                 throw new IllegalArgumentException("interval must be positive and non-null");
             }
-            long maxRangeIntervalMs = (long) monitor.dataPoints * monitor.scrapeInterval.toMillis();
+            long maxRangeIntervalMs = monitor.dataPoints * monitor.scrapeInterval.toMillis();
             if (interval.toMillis() > maxRangeIntervalMs) {
                 throw new IllegalArgumentException(
                         String.format(
